Rename the module.
