+++ ai/acme-askforhelp/.gitignore
