+++ ai/acme-askforhelp/src/main/java/com/microsoft/azure/acme/askforhelp/common/AzureOpenAIClient.java
@@ -1,4 +1,4 @@
-package com.microsoft.azure.acme.ai.common;
+package com.microsoft.azure.acme.askforhelp.common;
 
 import com.azure.ai.openai.OpenAIClient;
 import com.azure.ai.openai.models.ChatCompletions;
