+++ ai/acme-askforhelp/src/main/java/com/microsoft/azure/acme/askforhelp/common/TextSplitter.java
@@ -1,4 +1,4 @@
-package com.microsoft.azure.acme.ai.common;
+package com.microsoft.azure.acme.askforhelp.common;
 
 import com.knuddels.jtokkit.Encodings;
 import com.knuddels.jtokkit.api.Encoding;
