+++ ai/acme-askforhelp/src/main/java/com/microsoft/azure/acme/askforhelp/common/vectorstore/VectorStore.java
@@ -1,4 +1,4 @@
-package com.microsoft.azure.acme.ai.common.vectorstore;
+package com.microsoft.azure.acme.askforhelp.common.vectorstore;
 
 import java.util.List;
